refactor: consolidate proxy routes and remove unused image and raw routes
